+++ src/animation/helper.h
@@ -24,7 +24,7 @@ float map(float value, float inMin, float inMax, float outMin, float outMax);
 float dist(float x1, float y1, float x2, float y2);
 
 // Drawing Helpers
-void drawQuadraticBezier(LGFX_Sprite* sprite, float x0, float y0, float x1, float y1, float x2, float y2, uint32_t color);
-void fillQuadraticBezier(LGFX_Sprite* sprite, Point anchor, float x0, float y0, float x1, float y1, float x2, float y2, uint32_t color);
+void drawQuadraticBezier(LGFX_Sprite* sprite, float x0, float y0, float x1, float y1, float x2, float y2, uint16_t color);
+void fillQuadraticBezier(LGFX_Sprite* sprite, Point anchor, float x0, float y0, float x1, float y1, float x2, float y2, uint16_t color);
 
 float randomFloat(float minValue, float maxValue);
